添加内存分配宏定义 添加分配缓冲区 添加自定义异常 添加异常单元测试 添加分配缓冲区单元测试 基础服务改成每次循环处理内部消息池

+++ tcp_service.cc
@@ -137,12 +137,12 @@ void TcpService::ReadCallback(uv_stream_t * stream, ssize_t nread, const uv_buf_
     if (nread < 0)
     {
         if (nread != UV_EOF && service->_logger)
-            service->_logger->Error("%s Read Error: %s", service->GetName(), uv_strerror(nread));
+            service->_logger->Error("%s Read Error: %s", service->GetName(), uv_strerror((int)nread));
         
         service->Close();
     }
     else if (nread > 0)
-        service->OnRecv(buf->base, nread);
+        service->OnRecv(buf->base, (int)nread);
 
     free(buf->base);
 }
